+++ HISTORY.rst
@@ -11,6 +11,12 @@ TODO
 * filtering sample by status or error make a lot of queries. Consider writing
   coroutines or reading ValidationResult as pages
 
+0.3.1.dev0
+----------
+
+* Change ``Auth.__str__()``: now it returns ``Token for Foo Bar will expire in HH:MM:SS``
+* add ``Auth.get_domains`` which returns ``self.claims['domains']``
+
 0.3.0 (2020-01-14)
 ------------------